pyproject.toml
[project]
name = "devin-mcp-proxy"
version = "0.1.0"
description = "A FastMCP proxy server for Devin MCP API"
dependencies = [
    "fastmcp>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["main.py"]